Implement tenants.exists using new endpoint
