+++ .github/workflows/main.yaml
@@ -15,7 +15,7 @@ on:
 
 env:
   OLD_WEAVIATE_VERSION: 1.23.13
-  NEW_WEAVIATE_VERSION: 1.24.5
+  NEW_WEAVIATE_VERSION: 1.24.5-7fa1d5e
 
 jobs:
   lint-and-format:
